Move them to shared, not models
