+++ package.json
@@ -19,7 +19,7 @@
     "http-proxy": "^1.18.1",
     "jsonwebtoken": "^8.5.1",
     "jwks-rsa": "^1.8.1",
-    "models": "file:./models",
+    "shared": "file:./shared",
     "multer": "^1.4.2",
     "nodemon": "^2.0.4",
     "pg": "^7.3.0",
