+++ shared/models/FundingTime.ts
