+++ src/entity/FlattenedEnrollment.ts
@@ -1,6 +1,6 @@
 import { Entity, Column, PrimaryGeneratedColumn, ManyToOne } from 'typeorm';
 
-import { FlattenedEnrollment as FlattenedEnrollmentInterface } from 'models';
+import { FlattenedEnrollment as FlattenedEnrollmentInterface } from 'shared/models';
 
 import { EnrollmentReport } from './EnrollmentReport';
 import {
